requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
astronomer-cosmos>=1.10.1
apache-airflow-providers-postgres==6
apache-airflow-providers-common-sql>=1.20.0